Fix latency model bug (#734)
* Add failing test

* Fix int overflow in `LatencyModel`

+++ nautilus_trader/backtest/models.pyx
@@ -140,10 +140,10 @@ cdef class LatencyModel:
 
     def __init__(
         self,
-        int base_latency_nanos = NANOSECONDS_IN_MILLISECOND,
-        int insert_latency_nanos = 0,
-        int update_latency_nanos = 0,
-        int cancel_latency_nanos = 0,
+        uint64_t base_latency_nanos = NANOSECONDS_IN_MILLISECOND,
+        uint64_t insert_latency_nanos = 0,
+        uint64_t update_latency_nanos = 0,
+        uint64_t cancel_latency_nanos = 0,
     ):
         Condition.not_negative_int(base_latency_nanos, "base_latency_nanos")
         Condition.not_negative_int(insert_latency_nanos, "insert_latency_nanos")
